+++ app/Http/Controllers/PageController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\PostRequest;
 use App\Repositories\PageRepository;
+use Illuminate\Support\Facades\Gate;
 
 class PageController extends Controller
 {
@@ -17,12 +18,18 @@ class PageController extends Controller
 
     public function index()
     {
+        if(!Gate::allows('page.index')){
+            abort(401);
+        }
         $pages = $this->pageRepository->adminPaginate();
         return view('admin.page.index', compact('pages'));
     }
 
     public function create()
     {
+        if(!Gate::allows('page.create')){
+            abort(401);
+        }
         return view('admin.page.create');
     }
 
@@ -34,6 +41,9 @@ class PageController extends Controller
 
     public function edit($id)
     {
+        if(!Gate::allows('page.edit')){
+            abort(401);
+        }
         $page = $this->pageRepository->findById($id);
         return view('admin.page.edit', compact('page'));
     }
@@ -46,12 +56,15 @@ class PageController extends Controller
 
     public function update(PostRequest $request, $id)
     {
-        return $this->pageRepository->update($id, $request->except('_token'));
+        $this->pageRepository->update($id, $request->except('_token'));
         return redirect('admin/page');
     }
 
     public function destroy($id)
     {
+        if(!Gate::allows('page.destroy')){
+            abort(401);
+        }
         $this->pageRepository->delete($id);
         return redirect()->back()->with('success', '删除页面成功！');
     }
